break up GedDocumentMongo
Fixes #625 - codebeat complexity warnings

Break into pieces to keep each piece small and simple.
