+++ gedbrowser-mongo-dao/src/main/java/org/schoellerfamily/gedbrowser/persistence/mongo/domain/GedDocumentMongo.java
@@ -1,13 +1,6 @@
 package org.schoellerfamily.gedbrowser.persistence.mongo.domain;
 
-import java.util.ArrayList;
-import java.util.List;
-
 import org.schoellerfamily.gedbrowser.datamodel.GedObject;
-import org.schoellerfamily.gedbrowser.persistence.GedDocumentLoader;
-import org.schoellerfamily.gedbrowser.persistence.domain.GedDocument;
-import org.schoellerfamily.gedbrowser.persistence.mongo.domain.visitor.GedDocumentMongoVisitor;
-import org.schoellerfamily.gedbrowser.persistence.mongo.domain.visitor.TopLevelGedDocumentMongoVisitor;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.annotation.Transient;
 import org.springframework.data.mongodb.core.index.Indexed;
@@ -18,7 +11,7 @@ import org.springframework.data.mongodb.core.index.Indexed;
  * @param <G> the associated GedObject subclass
  */
 public abstract class GedDocumentMongo<G extends GedObject>
-        implements GedDocument<G> {
+        extends HasAttributes<G> implements Accepts {
     /** */
     @Id
     private String idString;
@@ -34,10 +27,6 @@ public abstract class GedDocumentMongo<G extends GedObject>
     /** */
     @Indexed
     private String dbName;
-
-    /** */
-    private final List<GedDocument<? extends GedObject>> attributes =
-            new ArrayList<>();
 
     /** */
     @Transient
@@ -122,58 +111,4 @@ public abstract class GedDocumentMongo<G extends GedObject>
     public final void setGedObject(final G gedObject) {
         this.gedObject = gedObject;
     }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public final List<GedDocument<? extends GedObject>> getAttributes() {
-        return attributes;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public final void setAttributes(
-            final List<GedDocument<? extends GedObject>> attributes) {
-        this.attributes.clear();
-        this.attributes.addAll(attributes);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public final void addAttribute(final GedDocument<?> attribute) {
-        attributes.add(attribute);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public final void clearAttributes() {
-        attributes.clear();
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public abstract void loadGedObject(GedDocumentLoader loader, GedObject ged);
-
-    /**
-     * Accept a visitor.
-     *
-     * @param visitor the visitor
-     */
-    public abstract void accept(TopLevelGedDocumentMongoVisitor visitor);
-
-    /**
-     * Accept a visitor.
-     *
-     * @param visitor the visitor
-     */
-    public abstract void accept(GedDocumentMongoVisitor visitor);
 }
